+++ src/main/java/game/map/editor/DrawTrianglesManager.java
@@ -137,24 +137,26 @@ public class DrawTrianglesManager
 
 			if (pointList.size() > 2) {
 				done = true;
-				success = true;
+				success = false;
 
 				Vector3f AB = GeometryUtils.getVectorBetween(pointList.get(0), pointList.get(1));
 				Vector3f AC = GeometryUtils.getVectorBetween(pointList.get(0), pointList.get(2));
 
 				Vector3f uAB = new Vector3f(AB);
-				Vector3f uAC = new Vector3f(AC);
 
 				uAB.normalize();
-				uAC.normalize();
 
 				Vector3f proj = uAB.scale(Vector3f.dot(uAB, AC));
 
 				Vector3f norm = Vector3f.sub(AC, proj);
 				if (norm.length() > 1e-4) {
 					norm.normalize();
-					new TriangleCutter(pointList.get(0), norm, editor.selectionManager.getSelectedObjects());
+					TriangleCutter cutter = new TriangleCutter(
+						pointList.get(0), norm, editor.selectionManager.getSelectedObjects());
+					success = cutter.wasSuccessful();
 				}
+				else
+					Logger.logWarning("Cut aborted: the side point is collinear with the cut line.");
 			}
 		}
 	}